requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
